+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * **(breaking-change)** Transfer plugin to southbridge (http://github.com/centosadmin)
 * Make preserve author of issue on copy
 * Add relation between Issue and WeeklySchedule 
+* Add issue associations to copy in settings
 
 # 0.0.1
 * Add ability to make issues copies