+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Make preserve author of issue on copy
 * Add relation between Issue and WeeklySchedule 
 * Add issue associations to copy in settings
+* If issue have due_date then copied issue have relative due_date 
 
 # 0.0.1
 * Add ability to make issues copies